// resources/js/types/app-settings.ts
export type AppSettings = {
    app_name: string;
    app_logo_url: string | null;
    app_tagline: string | null;
    app_description: string | null;
    repository_url: string | null;
    documentation_url: string | null;
    support_email: string | null;
    company_name: string | null;
    imprint_url: string | null;
    privacy_url: string | null;
};